+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## UNRELEASED
 
+- `Jsonnet.prototype.{extString,extCode,tlaString,tlaCode}` now accept an object to assign multiple variables at once.
+
 ## v3.1.1 (2026-04-02)
 
 - No changes since v3.1.0 -- v3.1.0 was not properly packaged during the release workflow.